net/devif_callback: change flags type from uint16_t to uint32_t
increase the size of the flag to prepare for precise xxx_POLL.

+++ net/bluetooth/bluetooth.h
@@ -254,9 +254,9 @@ FAR struct bluetooth_conn_s *
  *
  ****************************************************************************/
 
-uint16_t bluetooth_callback(FAR struct radio_driver_s *radio,
-                             FAR struct bluetooth_conn_s *conn,
-                             uint16_t flags);
+uint32_t bluetooth_callback(FAR struct radio_driver_s *radio,
+                            FAR struct bluetooth_conn_s *conn,
+                            uint32_t flags);
 
 /****************************************************************************
  * Name: bluetooth_recvmsg
