+++ net/can/can_sendmsg.c
@@ -79,8 +79,8 @@ struct send_s
  * Name: psock_send_eventhandler
  ****************************************************************************/
 
-static uint16_t psock_send_eventhandler(FAR struct net_driver_s *dev,
-                                        FAR void *pvpriv, uint16_t flags)
+static uint32_t psock_send_eventhandler(FAR struct net_driver_s *dev,
+                                        FAR void *pvpriv, uint32_t flags)
 {
   FAR struct send_s *pstate = pvpriv;
